pyproject.toml
[tool.poetry]
name = "maid-api"
version = "0.1.0"
description = ""
authors = ["mc_fdc <Masato.04.11.2007@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.104.1"
uvicorn = {extras = ["standard"], version = "^0.23.2"}
python-jose = "^3.3.0"
aiohttp = "^3.8.6"
aiomysql = "^0.2.0"
python-dotenv = "^1.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
